% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_definitions.R
\name{get_definitions}
\alias{get_definitions}
\title{Finds the definition(s) for a given acronym}
\usage{
get_definitions(acronym)
}
\arguments{
\item{acronym}{the computer acronym}
}
\value{
a data frame with the definitions
}
\description{
Finds the definition(s) for a given acronym
}
\examples{
defs <- get_definitions("AAP")
defs
defs$definition
# [1] "Address Allocation Protocol (Multicast)"
# [2] "Applications Access Point"
# [3] "Association of American Publishers (Org., USA)"
}
