% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vera.R
\docType{package}
\name{vera}
\alias{vera}
\alias{vera-package}
\title{The VERA computer acronyms list}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 12006 rows and 2 columns.}
\usage{
vera
}
\description{
This package contains the complete list
of computer acronyms from the "Virtual Entity of Relevant Acronyms" (VERA).
}
\details{
The original source is at: https://www.ohei2.de/vera/vueber-e.html
}
\keyword{datasets}
