% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vera.R
\name{lsa}
\alias{lsa}
\title{Perform local sensitivity analysis}
\usage{
lsa(mod, fun, par, var, eps = 1e-08, ...)
}
\arguments{
\item{mod}{a mrgsolve model object}

\item{fun}{generating simulated for sensitivity analysis (see details)}

\item{par}{parameter names as character vector or comma-separated string}

\item{var}{output names (compartment or capture) as character vector
or comma-separated string}

\item{eps}{parameter change value for sensitivity analysis}

\item{...}{arguments passed to \code{fun}}
}
\description{
Perform local sensitivity analysis
}
